Add scrolling on channels lists
Problem: when there are too many channels not all of them are
accessible because the late elements of list are hidden below viewport.

Solution: add possibility to scroll the channels list

Done: The list of channels has been made scrollable (actually wrapped
with <div> with overflow-y: auto and now it user can scroll through the
list of channels. Scroll only works for the list channels making locale
select and "new channel" button always visible.

Padding-right was removed from nav-drawer for the mobile version of the
website, because it made the scrollable list look pretty ugly.

NOTE: needs checks in firefox as well as in safari and (OMG, IE/Edge).

If ok, this should resolve #37