[#35] Ellipsis channel name in new-post
Fix .bottom layout for different width channel name and post-button
title. Now post button takes as much space as it needs and channel-name
takes the rest of the space in line.

Described fix allowed to ellipsis channel name when it is too long
The title attribute on span has been also added to display the whole
name on hover;

Partially resolves #35